added conditional REST calls

+++ desora.ro/js/app.js
@@ -9,6 +9,7 @@ class App {
 
 		let alarme = '';
 		let alarmeData;
+		let avemDatabase = false;
 
 		let ceas = new Ceas();
 		ceas.adaugaSegmente();
@@ -69,8 +70,10 @@ class App {
 			// Hack: daca lansam applicatia de pe alt site decat cel cu db-ul incarca un json local
 			if (window.location.host == 'www.desora.ro') {
 				alarmeData = "alarme-data";
+				avemDatabase = true;
 			} else {
 				alarmeData = "alarmeData.json";
+				avemDatabase = false;
 			}
 
 			// luam lista json de la server cu alarme
@@ -125,11 +128,13 @@ class App {
 					var randNou = tabel.find('tr.hide').clone(true).removeClass('hide table-line');
 					tabel.find('table').append(randNou);
 
-					$.post("alarme-data/insert", {
-						operation: "insert"
-					}, function (data, status) {
-						randNou[0].children[0].innerText = JSON.parse(data).insertId;
-					});
+					if (avemDatabase) {
+						$.post("alarme-data/insert", {
+							operation: "insert"
+						}, function (data, status) {
+							randNou[0].children[0].innerText = JSON.parse(data).insertId;
+						});
+					}
 
 				});
 
@@ -138,14 +143,15 @@ class App {
 					// console.log(idDeSters);
 
 					let url = 'alarme-data/delete';
-					$.ajax({
-						url: url + '?' + $.param({
-							"id": idDeSters
-						}),
-						type: 'DELETE',
-					});
-
-					$(this).parents('tr').detach();
+					if (avemDatabase) {
+						$.ajax({
+							url: url + '?' + $.param({
+								"id": idDeSters
+							}),
+							type: 'DELETE',
+						});
+						$(this).parents('tr').detach();
+					}
 				});
 
 				$("table tr").click(function (rand) {
@@ -166,18 +172,20 @@ class App {
 					$("#secunde").val(secunde);
 
 					let url = 'alarme-data/update';
-					$.ajax({
-							url: url + '?' + $.param({
-								"id": id,
-								"nume": nume,
-								"ore": ore,
-								"minute": minute,
-								"secunde": secunde
-							}),
-							type: 'PUT',
-						}).done(function () { /* console.info( "success update "); */ })
-						.fail(function () { /* console.warn( "error update" ); */ })
-						.always(function () { /*console.log( "complete update"); */ });
+					if (avemDatabase) {
+						$.ajax({
+								url: url + '?' + $.param({
+									"id": id,
+									"nume": nume,
+									"ore": ore,
+									"minute": minute,
+									"secunde": secunde
+								}),
+								type: 'PUT',
+							}).done(function () { /* console.info( "success update "); */ })
+							.fail(function () { /* console.warn( "error update" ); */ })
+							.always(function () { /*console.log( "complete update"); */ });
+					}
 				});
 			});
 		});
